refactor(calendar): extract context label util
