+++ src/components/calendar/index.ts
@@ -5,4 +5,5 @@ export { CalendarModule } from "./calendar-module";
 
 // Re-export all calendar components
 export * from "./components";
+export * from "./utils/context-label";
 export * from "./views";
